RoomSync Backend main/requirements.txt
Django>=5.0.0
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
